feat(stream): support int/float dtypes for streams, default float32
Streams were pinned to float32 (hardcoded in LSLSource), even though the buffers
and Zarr recording already honor StreamInfo.dtype. Open it up:

- StreamInfo: add SUPPORTED_DTYPES (float16/32/64, int8/16/32/64 — every LSL
  wire format plus float16 as a storage cast; no unsigned, none on the LSL
  wire). __post_init__ normalises str/type/np.dtype and validates. Default
  stays float32.
- LSLSource(stream_name, dtype='float32'): cast incoming samples to the chosen
  dtype; dtype=None honors the outlet's native wire format (lossless for int
  amps). Fixes the old connect() docstring that claimed dtype came from
  metadata while hardcoding float32.
- The window handed to @pipeline.extract is ALWAYS upcast to float32 — at the
  live boundary (Stream.get_window) and the training boundary (iter_labeled_
  windows / iter_aligned_windows) — so feature/ML code never branches on dtype,
  while buffers + recording stay compact (e.g. int16 = half memory/disk).

Net win: 16-bit EMG amps can stream/record losslessly at half the size.
Recording already used info.dtype, so no change needed there.

Also: bare '# type: ignore' on the reconnect() call so ty honors it.

+++ myogestic/session/_windows.py
@@ -51,7 +51,7 @@ def iter_labeled_windows(
         if win_samples < 1:
             continue
 
-        data = np.array(sess.stores[stream_name])
+        data = np.array(sess.stores[stream_name]).astype(np.float32, copy=False)
         ts = np.array(sess.ts_stores[stream_name])
         events = sess.label_track
         if len(data) == 0 or not events:
@@ -125,9 +125,12 @@ def iter_aligned_windows(
         if win_samples < 1:
             continue
 
-        primary_data = np.array(sess.stores[primary_stream])
+        primary_data = np.array(sess.stores[primary_stream]).astype(np.float32, copy=False)
         primary_ts = np.array(sess.ts_stores[primary_stream])
-        aligned_data = {name: np.array(sess.stores[name]) for name in aligned_streams}
+        aligned_data = {
+            name: np.array(sess.stores[name]).astype(np.float32, copy=False)
+            for name in aligned_streams
+        }
         aligned_ts = {name: np.array(sess.ts_stores[name]) for name in aligned_streams}
 
         if (
